pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qmd-search"
version = "0.1.0"
description = "QMD Local Document Search – Query → Model → Documents"
requires-python = ">=3.11"
dependencies = [
    "click==8.1.7",
    "pyyaml==6.0.2",
]

[project.optional-dependencies]
bm25 = ["rank-bm25==0.2.2"]
embed = [
    "sentence-transformers==3.0.1",
    "torch>=2.0.0",
]
llm = ["openai>=1.0.0", "anthropic>=0.40.0", "python-dotenv>=1.0.0"]
ui = ["streamlit>=1.30.0"]
dev = ["pytest==8.3.3", "pytest-cov==5.0.0"]

[project.scripts]
qmd = "qmd.cli:main"

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
